+++ packages/expo-router/src/layouts/StackClient.tsx
@@ -20,9 +20,7 @@ import {
 } from '../react-navigation/native';
 import type { NativeStackNavigationOptions } from '../react-navigation/native-stack';
 import type { NativeStackNavigationConfig } from '../react-navigation/native-stack/types';
-import { makePopAction } from '../react-navigation/native-stack/utils/makePopAction';
 import { IsWithinNativeNavigator, unstable_integrateWithRouter } from '../standard-navigation';
-import { subscribePopToTopOnParentTabPress } from '../standard-navigation/subscribePopToTopOnParentTabPress';
 import { isChildOfType } from '../utils/children';
 import { Protected } from '../views/Protected';
 import { StackRouter } from './stack-router';
@@ -34,6 +32,7 @@ import {
   StackTitle,
   StackToolbar,
   appendScreenStackPropsToOptions,
+  createBaseStackProps,
   mapProtectedScreen,
   validateStackPresentation,
 } from './stack-utils';
@@ -46,19 +45,10 @@ const RNStack = unstable_integrateWithRouter<
   StackRouterOptions,
   NativeStackNavigatorCreateProps
 >(createStandardNativeStackNavigator, StackRouter, {
-  createProps: ({
-    state,
-    dispatch,
-    dispatchSync,
-    navigation,
-    isPreloaded,
-    isRemovalPrevented,
-  }) => ({
-    isPreloaded,
-    isRemovalPrevented,
-    pop: makePopAction(dispatchSync, state.key),
-    removeRoutes: (routeNames) => dispatch({ type: 'REMOVE_ROUTES', payload: { routeNames } }),
-    subscribePopToTopOnParentTabPress: () => subscribePopToTopOnParentTabPress(navigation, state),
+  createProps: (args) => ({
+    ...createBaseStackProps(args),
+    removeRoutes: (routeNames) =>
+      args.dispatch({ type: 'REMOVE_ROUTES', payload: { routeNames } }),
   }),
 });